test: missing test cases
