+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/set/multiset/const/ConstMultiSetImpl.kt
@@ -1,4 +1,4 @@
 package xyz.lbres.kotlinutils.set.multiset.const
 
-// final implementation of ConstMultiSet class
+// final implementation of ConstMultiSet
 internal class ConstMultiSetImpl<E>(initialElements: Collection<E>) : ConstMultiSet<E>(initialElements)
