+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/set/multiset/const/ConstMutableMultiSetImpl.kt
@@ -1,4 +1,4 @@
 package xyz.lbres.kotlinutils.set.multiset.const
 
-// final implementation of ConstMutableMultiSet class
+// final implementation of ConstMutableMultiSet
 class ConstMutableMultiSetImpl<E>(initialElements: Collection<E>) : ConstMutableMultiSet<E>(initialElements)
