+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/const/ConstMutableMultiSetTest.kt
@@ -1,5 +1,6 @@
 package xyz.lbres.kotlinutils.set.multiset.const
 
+import xyz.lbres.kotlinutils.internal.constants.Suppressions
 import xyz.lbres.kotlinutils.list.IntList
 import xyz.lbres.kotlinutils.set.multiset.impl.MultiSetImpl
 import xyz.lbres.kotlinutils.set.multiset.testutils.* // ktlint-disable no-wildcard-imports no-unused-imports
@@ -7,10 +8,12 @@ import kotlin.test.Test
 
 class ConstMutableMultiSetTest {
     private fun <T> createSet(): (Collection<T>) -> ConstMutableMultiSet<T> = { ConstMutableMultiSetImpl(it) }
+    private fun <T> createOtherSet(): (Collection<T>) -> MultiSetImpl<T> = { MultiSetImpl(it) }
 
     @Test
     fun testConstructor() {
         fun <T> runTest(map: Map<String, Any>) {
+            @Suppress(Suppressions.UNCHECKED_CAST)
             val values: Collection<T> = map["values"] as Collection<T>
             val set: ConstMutableMultiSet<T> = ConstMutableMultiSetImpl(values)
             testConstructed(set, map)
@@ -22,7 +25,7 @@ class ConstMutableMultiSetTest {
         multiSetConstructorCompListTestValues.forEach { runTest<List<Comparable<*>>>(it) }
     }
 
-    @Test fun testEquals() = runEqualsTests(createSet(), createSet(), createSet()) { MultiSetImpl(it) }
+    @Test fun testEquals() = runMutableEqualsTests(createSet(), createSet(), createSet(), createOtherSet())
 
     @Test fun testContains() = runMutableContainsTests(createSet(), createSet(), createSet())
     @Test fun testContainsAll() = runMutableContainsAllTests(createSet())
@@ -36,23 +39,22 @@ class ConstMutableMultiSetTest {
 
     @Test
     fun testMinus() {
-        runMinusTests(createSet(), createSet(), createSet(), createSet(), createSet()) { MultiSetImpl(it) }
+        runMinusTests(createSet(), createSet(), createSet(), createSet(), createOtherSet())
     }
 
     @Test
     fun testPlus() {
-        runPlusTests(createSet(), createSet(), createSet(), createSet(), createSet()) { MultiSetImpl(it) }
+        runPlusTests(createSet(), createSet(), createSet(), createSet(), createOtherSet())
     }
 
     @Test
     fun testIntersect() {
-        runIntersectTests(createSet(), createSet(), createSet()) { MultiSetImpl(it) }
+        runIntersectTests(createSet(), createSet(), createSet(), createOtherSet())
     }
 
     @Test fun testIsEmpty() = runMutableIsEmptyTests(createSet(), createSet())
     @Test fun testGetCountOf() = runMutableGetCountOfTests(createSet(), createSet())
 
     @Test fun testIterator() = runMutableIteratorTests(createSet(), createSet())
-
     @Test fun testToString() = runMutableToStringTests(createSet(), createSet())
 }
